rootston: split rendering code into render.c
